Moved syscalls to separate files to fix aliasing problems.
