+++ libgloss/riscv/sys_chdir.c
@@ -1,0 +1,13 @@
+#include <machine/syscall.h>
+
+//----------------------------------------------------------------------
+// chdir
+//----------------------------------------------------------------------
+// Stub.
+
+int
+_chdir(const char *path)
+{
+  return -1;
+}
+
