+++ libgloss/riscv/sys_chmod.c
@@ -1,0 +1,14 @@
+#include <machine/syscall.h>
+
+//----------------------------------------------------------------------
+// chmod
+//----------------------------------------------------------------------
+// Stub.
+
+int
+_chmod(const char *path, mode_t mode)
+{
+  return -1;
+}
+
+
