+++ libgloss/riscv/sys_chown.c
@@ -1,0 +1,12 @@
+#include <machine/syscall.h>
+
+//----------------------------------------------------------------------
+// chown
+//----------------------------------------------------------------------
+// Stub.
+
+int _chown(const char *path, uid_t owner, gid_t group)
+{
+  return -1;
+}
+
