+++ libgloss/riscv/sys_faccessat.c
@@ -1,0 +1,12 @@
+#include <machine/syscall.h>
+
+//------------------------------------------------------------------------
+// faccessat
+//------------------------------------------------------------------------
+// Permissions of a file (by name) in a given directory.
+
+int _faccessat(int dirfd, const char *file, int mode, int flags)
+{
+  return syscall_errno (SYS_faccessat, dirfd, file, mode, flags, 0, 0);
+}
+
