+++ libgloss/riscv/sys_getcwd.c
@@ -1,0 +1,13 @@
+#include <machine/syscall.h>
+
+//----------------------------------------------------------------------
+// getcwd
+//----------------------------------------------------------------------
+// Stub.
+
+char *
+_getcwd(char *buf, size_t size)
+{
+  return NULL;
+}
+
