+++ libgloss/riscv/sys_kill.c
@@ -1,0 +1,15 @@
+#include <machine/syscall.h>
+
+//------------------------------------------------------------------------
+// kill
+//------------------------------------------------------------------------
+// Send a signal. Minimal implementation for a system without processes
+// just causes an error.
+
+int
+_kill(int pid, int sig)
+{
+  errno = EINVAL;
+  return -1;
+}
+
