+++ libgloss/riscv/sys_sysconf.c
@@ -1,0 +1,19 @@
+#include <machine/syscall.h>
+
+//----------------------------------------------------------------------
+// sysconf
+//----------------------------------------------------------------------
+// Get configurable system variables
+
+long
+_sysconf(int name)
+{
+  switch (name)
+    {
+    case _SC_CLK_TCK:
+      return CLOCKS_PER_SEC;
+    }
+
+  return -1;
+}
+
