+++ libgloss/riscv/sys_utime.c
@@ -1,0 +1,13 @@
+#include <machine/syscall.h>
+
+//----------------------------------------------------------------------
+// utime
+//----------------------------------------------------------------------
+// Stub.
+
+int
+_utime(const char *path, const struct utimbuf *times)
+{
+  return -1;
+}
+
